+++ postgresql/main.go
@@ -22,7 +22,7 @@ type Postgresql interface {
 // ----------------------------------------------------------------------------
 
 // Identfier of the  package found messages having the format "senzing-6423xxxx".
-const ProductId = 6423
+const ComponentId = 6423
 
 // ----------------------------------------------------------------------------
 // Variables